Return mock from init if set in config
